migrated tests to junit 5
