+++ src/test/java/com/github/flo456123/common/parser/ParserTest.java
@@ -3,20 +3,19 @@ package com.github.flo456123.common.parser;
 import com.github.flo456123.common.element.Element;
 import com.github.flo456123.common.element.ElementFactory;
 import com.github.flo456123.common.types.Substance;
-import org.junit.Test;
+import org.junit.jupiter.api.Test;
 
-import static org.junit.Assert.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertEquals;
 
 public class ParserTest {
 
-    // Substance parser test
     @Test
     public void validSubstanceTest1() {
-        Substance actual = Parser.parseSubstanceString("2HCl");
-
         Element elementReactant1 = ElementFactory.createElement("H", 1);
         Element elementReactant2 = ElementFactory.createElement("Cl", 1);
         Substance expected = new Substance(2, elementReactant1, elementReactant2);
+
+        Substance actual = Parser.parseSubstanceString("2HCl");
 
         assertEquals(expected.n(), actual.n());
         assertEquals(expected.elements(), actual.elements());
@@ -24,11 +23,11 @@ public class ParserTest {
 
     @Test
     public void validSubstanceTest2() {
-        Substance actual = Parser.parseSubstanceString("2NaCl_2");
-
         Element elementReactant1 = ElementFactory.createElement("Na", 1);
         Element elementReactant2 = ElementFactory.createElement("Cl", 2);
         Substance expected = new Substance(2, elementReactant1, elementReactant2);
+
+        Substance actual = Parser.parseSubstanceString("2NaCl_2");
 
         assertEquals(expected.n(), actual.n());
         assertEquals(expected.elements(), actual.elements());
@@ -36,11 +35,11 @@ public class ParserTest {
 
     @Test
     public void validSubstanceTest3() {
-        Substance actual = Parser.parseSubstanceString("HS_4");
-
         Element elementProduct3 = ElementFactory.createElement("H", 1);
         Element elementProduct4 = ElementFactory.createElement("S", 4);
         Substance expected = new Substance(1, elementProduct3, elementProduct4);
+
+        Substance actual = Parser.parseSubstanceString("HS_4");
 
         assertEquals(expected.n(), actual.n());
         assertEquals(expected.elements(), actual.elements());
@@ -50,57 +49,70 @@ public class ParserTest {
     @Test
     public void countElementTest1() {
         int expected = 3;
+
         String elementString = "HClO_3";
         int actual = Parser.countElements(elementString);
+
         assertEquals(expected, actual);
     }
 
     @Test
     public void countElementTest2() {
         int expected = 2;
+
         String elementString = "HCl";
         int actual = Parser.countElements(elementString);
+
         assertEquals(expected, actual);
     }
 
     @Test
     public void countElementTest3() {
         int expected = 2;
+
         String elementString = "H_2O";
         int actual = Parser.countElements(elementString);
+
         assertEquals(expected, actual);
     }
 
     @Test
     public void countInvalidElementTest1() {
         int expected = 0;
+
         String elementString = "";
         int actual = Parser.countElements(elementString);
+
         assertEquals(expected, actual);
     }
 
-    // Atom parser test
     @Test
     public void parseAtomsTest1() {
         int expected = 2;
+
         String elementString = "H_2";
         int actual = Parser.parseAtoms(elementString);
+
         assertEquals(expected, actual);
     }
 
     @Test
     public void parseAtomsTest2() {
         int expected = 5;
+
         String elementString = "Na_5";
         int actual = Parser.parseAtoms(elementString);
+
         assertEquals(expected, actual);
     }
 
     @Test
     public void parseAtomsTest3() {
         int expected = 7;
+
         String elementString = "Cl_7";
         int actual = Parser.parseAtoms(elementString);
+
         assertEquals(actual, expected);
     }
